moved around components to better match directory structure
