+++ frontend/src/components/dashboard/MovieTab.tsx
