+++ frontend/src/components/dashboard/MovieTable.tsx
